Additional visuals stories
Created new visuals stories in the backlog so all line items in this table have at least a corresponding GitHub Story they can link to. Some of these stories do not have a wireframe story as of yet, but when they do, the description text within each story will be updated with a link to the corresponding wireframe story.